requirements.txt
PyYAML==6.0
spotipy==2.23.0